Initial unittest for VenueServer.  No external connections are made for this level of test.  This test won't be added to unittest_all yet because it still requires the kill command (unix: cntrl-Z, "kill %") when finished to clean up the server.
Note that PyUnit (and JUnit as well) execute setUp() and tearDown() once for EACH test and make it difficult to setUp and tearDown something across multiple tests.  This is not ideal for our server tests, but we work around it to keep a server running for multiple tests.

Testing:
   AddVenue, RemoveVenue, Checkpoint, AddAdministrator, RemoveAdministrator, GetAdministrator, GetSetStorageLocaion, GetSetEncryptAllMedia
Not Testing yet:
   Persistence, SetDefaultVenueServer