Allow AGNodeService and AGServiceManager to shutdown cleanly without needing to call os._exit().
The main problem was that since we are blocking when listening with
handle_request() (hosting/pyGlobus/Server.py), the Server's listening
thread would not exit.

By closing the server in Server's Stop() method, the thread will stop
listening and be able to shut itself down.  We also catch the exception that
closing the server causes the blocking listen() to throw ( from inside
handle_request() ).

Also added a Stop() method for AGNodeService to shutdown its internal
ServicePackageRepository.