Fixed so right venue in the list box gets selected after adding a venue. bugzilla #237 When changing to a sorted list box the venue doesn't get added last and it is then wrong to always select the last item in the list.
2 venues with the same name can not be added, or a venue can not be modified to have the same name
as another venue. I had to do this because the only way to find out the position of a venue is to
search for the name in the listbox, 2 venues with the same name could result in wrong venue selected.

Let me know if I should do this some other way or if it is correct to not allow 2 venues with the same name.
----------------------------------------------------------------------