Change the vic "quality" default setting from 85 to 75, this allowed a tranmitting vic (VideoProducer) to startup and send a camera signal in linux on ag2-test without segfaulting.
  Note the picture still seemed pixelated when I unmuted it (didn't wait
long though), due to the vic version, but at least it didn't crash on
startup.

I tested it on windows also and video also came up fine -- this seems
like a very safe change to me, but if this could have any repercussions,
feel free change it back and we can test settings a little more.

Settings at 75 and under seemed to work in my case.  Maybe it should be set
 even lower just in case?