Asynchronous IO mods, service side.
These are duplicates of EventService.py and TextService.py, modified
for asynch IO. If we deicde that we want to switch completely to
asynch IO, these can be renamed to the other names. If we want to keep
both flavors around, we need to do some factoring. There is common code
in the threaded and asynch versions of these.