Added Administrate Roles button to VenueClient "Venue" menu.
The button should be grayed out when you are unable to Administrate
the venue. (I can't verify this entirely on my own yet since I can't revoke
my own administration rights, but it is grayed out when I'm not connected to
venues).

To do this needed to add a DetermineRoles soap function to venue so users
can tell what roles they are in.

I wanted to the VenueClient button to modify the entire venue (from
VenueManagement), but that requires knowledge of the VenueServer (possible),
and currently relies on the base VenueManagement Frame being set up.  It
will take a little bit of work to get that working if we want it (for
next release).