Initialize Globus stuff before trying to start personal node, as the personal node communications requires a valid cert.
(addressing that issue is one that we're not addressing at this point,
but it probably be should. For intra-node inter-machien comms, we don't
have to have credentials that are known oustide the node. Bootstrapping
the credentials onto the components of a multi-machien node might
be hard, however; worth looking into).