Added DataService class. The DataService has two SOAP methods, RegisterVenue and UnRegisterVenue, used by the VenueServer to create a venue specific data store. The RegisterVenue call creates a DataStore that uses the DataService transfer engine for data upload and download and creates an Event client to send data events to the venue. The venue can get datadescriptions etc from the DataStore SOAP interface.  When a venue is shut down, UnRegisterVenue is called to shut down the DataStore SOAP interface, eventclient etc.
DataDescriptionContainer doesn't send PersonalDataEvents any more. Before, personal data stores used an eventclient and venue data stores used the eventservice,  but now both uses an EventClient so they send the same events.

Persistence is moved from VenueServer to DataStore. When a datastore is shut down, it stores the datadescriptions in a config file.

LEFT TO DO:
 Be able to ctrl-c the dataservice and shut down cleanly
 Ideally if the dataservice would go down it should save its current datastores info (url - venue id) in a config file that can be loaded when restarted.