Removed data persistence, now in datastore Removed transferengine, now in dataservice
The VenueServer config file now has an option for setting DataServiceUrl to create a remote datastore. If DataServiceUrl = '' the VenueServer creates its own DataService.

When the venue server creates a new venue, or loads it from persistence file, it calls DataService.RegisterVenue with event and venue information and adds the datastore to the venue. The venue will receive events from the datastore and can use its soap interface to get info. IF THE DATASERVICE IS NOT RUNNING WHEN WE ADD A VENUE OR START UP THE SERVER, THE VENUE WILL NOT HAVE A DATA STORE (NEEDS SOME MORE WORK).