Add method to get version from services remotely (bug 271):
GetAGTkVersion() is a soap method that returns a string for all services
that inherit ServiceBase.  If a different method name is preferred, and
 everyone agrees on one, it can easily be changed.

This at least includes and has been tested with VenueServer, Venue,
VenueClient, DataService, NodeService, and ServiceManager.

I wasn't able to get it to work with an individual node services (i.e.
VideoConsumerService) that are printed out when personalNode starts,
but I don't know why yet.

For a simple test, test_ServiceVersion.py takes urls as its argument list and
simply calls the soap method on them and prints the version.