Remove functions that were not being used (and were broken).
Venue.AddAdministrator, RemoveAdministrator, SetAdministrator,
and GetAdministrators are no longer being used.

All these can be done with the more generic roles interface, which
is what we have been using since 2.1.

If anyone would like me to add these old functions back in and convert them
to call the role functions, i can do that fairly quickly, but this
is just more code to test that is not called (although maybe it could
be called if anyone tried to use 2.0 venuemanagement with a 2.1.2 server).

They have not been used in a release since at least 2.0 because the
administrators variable they are referencing didn't exist in 2.1.