Catch GSITCP exceptions on venue entry; if they are authorization failures, use GetSNTPTime() to check the local client clock against the timeserver (currently wired to ntp-1.accessgrid.org).
If there is a significant (arbitrarily defined currently as
ten seconds) difference between the local and NTP clocks,
flag this as an error to the user.

Note that the user, even with a correct clock, could see these errors
if the server clock is off. It may be useful to do this check
on server startup as well.