Help prevent popups from affecting the url of the shared browser.
Events don't have enough information to tell if an event is from a
popup or a url the user clicked.  So, we wait until a page finishes
loading before navigate events (which could be popups or a user-click)
can start to load a new page.

It's a workaround that we can remove when the apis allow us to get more
information from the browser.

To detect when the main page is finished loading we have to ignore
navigation events while we're loading the main page and also
ignore DocumentComplete events for popups (urls that started loading
after the main page started loading).

The changes may not fix 100% of the cases, but it correctly loads the
three sites I could find with popups.