Modifications to allow command-line usage.
You can now run 'certmgr.py command'  to run single command as if
you were within the interactive certmgr.

I have extended this to take a DN in addition to a certificate number in the
places where the interface expects a cert number.

To specify ID or CA mode for the command, use the --id or --ca argument.

To force an overwrite of an existing file on export, use -f or --force.

Be careful: deletes from the command line do not ask for confirmation.