In fixing bug 975, I got an error when highlighting a role to remove. Action.RemoveRole now looks at the names of the roles to determine which  one to remove.
The problem was different instances of Role were being compared for
equality to find which one to remove.
Since objects are sent over soappy, it seems difficult to make sure
all Actions contain references to a single Role instead of similar
instances of them.  If that's desired, we can look into it, but at the
moment, an Action's removal of a Role simply uses the "name" attribute to
testing for equality.

Overriding the __eq__ also seemed like a less attractive alternative,
which was confirmed to be worse when it broke the standard hash functions.