Fix problem with VenueServer persistence and shared application data.
Endlines in shared app data prevented VenueServer from loading the .dat
file.  Simple encoding: replace "_" with "__" and replace endlines with
an underscore (and a character on each side of it so string.replace works).
Note: If we used a regular expression to detect the single underscore, we
would use a little less storage space (I can look that up, or anyone who
knows how to do that offhand can do it, but this works for now).

Also make all data keys lowercase before storing and retrieving data since
 we read all keys as lowercase when loading our .dat files.

Tested with data from AGBasicImage.