Modifications to get help Shared Presentation on linux work.
1. importing GetPaths.py fails because the file doesn't exist
Modification involves first checking if either 'ooffice' or 'soffice' is in
the search path of the PATH env variable, if unsuccessful, only then try to
import GetPaths.

2.  For SharedPresentation.py to work with ImpressViewer, viewer.win needs
to be non-None
So in ImpressViewer.py I set 'self.win = 1'
 - Submitted by Doug Kosovic

I also doubled the connection timeout when making initial connection to
openoffice -- this is because openoffice can take a while to start on
older machines. --Eric